Move Syncable conformance to separate file
Extensions implementing Syncable for EventObject, SessionObject, UserActivity, and MetricsObject were moved from Syncable.swift to a new SyncableConformance.swift file for better separation of protocol definition and implementation.
